requirements.txt
pandas
requests
numpy
python-dotenv
jupyter
scikit-learn
matplotlib
xgboost
fastapi
uvicorn
jinja2
python-multipart	
mysql-connector-python
watchdog
joblib
ta
VaderSentiment
websockets
torch
transformers
shap
catboost
pymysql
yfinance
SQLAlchemy
textblob
lightgbm
pandas_ta
seaborn
imblearn
beautifulsoup4
scipy
schedule
starlette
python-dotenv
bs4
re
tenacity
concurrent.futures
ccxt
polars
urllib3
nltk
unidecode
punkt_tab
pyarrow